requirements.txt
torch>=1.10.0
torchvision>=0.11.0
facenet-pytorch>=2.5.0
opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.4.0
sqlite3>=3.35.0